JavaProgrammingMasterClass/Section7/Challenges/Inheritance/src/main/java/org/example/Employee.java
package org.example;

import java.time.LocalDate;
import java.time.format.DateTimeFormatter;

public class Employee extends Worker {

    final DateTimeFormatter formatter = new Helper().formatter;
    private long employeeId;
    private LocalDate hireDate;

    public long getEmployeeId() {
        return employeeId;
    }

    public void setEmployeeId(long employeeId) {
        this.employeeId = employeeId;
    }

    public LocalDate getHireDate() {
        return hireDate;
    }

    public void setHireDate(LocalDate hireDate) {
        this.hireDate = hireDate;
    }

    public Employee() {}

    public Employee(String name, String birthDate, String endDate, long employeeId, String hireDate) {
        super(name, birthDate, endDate);
        this.employeeId = employeeId;
        this.hireDate = LocalDate.parse(hireDate, formatter);
    }

    @Override
    public String toString() {
        return "Employee{" +
                "employeeId=" + employeeId +
                ", hireDate='" + hireDate + '\'' +
                "} " + super.toString();
    }
}
